greatly simplify syncing of indexed-but-moved dirs
Personal benchmarks show a very significant performance gain when
calling get_path(). This is because we are no longer scanning all the
contents of every indexed-but-moved directory in _sync_all(), but rather
intelligently appending the new paths of the directories to check for
dirtiness to a deque used in the body of _sync_all().

Also fixes a few bugs (sorry, couldn't split this commit up):
- Indexing symlinks always indexes the real path that symlink refers to
- Fixes bug with the way new paths were calculated in recursive moves