Make preferred_dir content manager trait (#983)
* fix terminal test on Windows

* Make preferred_dir content manager trait

Move the preferred_dir trait to content manager, but keep the old one for backwards compatibility. The new location should also cause the value to be read later in the init, allowing us to avoid the deferred validation. Also fixes an issue with escaping the root dir on Windows if an absolute path is passed.

* Remove leading "/" from preferred_dir

Also validates set value to strip any leading slashes

* Allow filemanager to have OS path set

Will transparently convert to API path in validator.

* fix mypy

* fix ruff