Extend the existing gateway test to include non-local resources.
This change extends the mock kernelspec used for testing the gateway
client logic so that it includes one entry in the `resources` object
that is not a local file path relative to the `/kernelspecs/`
directory.

Adding that change is sufficient to make the tests fail with
the previous version of the code and to demonstrate that the change
in this PR fixes them.