Update to reflect `kernel_ws_protocol` being moved into BaseKernelWebsocketConnection
This change removes a temporary work-around to the required `kernel_ws_protocol`
field not being defined in the `BaseKernelWebsocketConnection` base class.

Previously, that field was only defined in the ZMQChannelsWebsocketConnection class,
so we had to reproduce it in the `GatewayWebsocketConnection` class in order to
get that class to work correctly. However, we only provided a minimal default value
of `None` rather than properly supporting the user setting their preferred websocket
protocol.

This workaround was made unnecessary by #1264 which moved the `kernel_ws_protocol`
field from the ZMQChannelsWebsocketConnection class to the base class.

Accordingly, this commit removes that temporary work-around that is no longer needed.
We needed to provide a value for that,