Run Gateway token renewers even if the auth token is empty.
Prior to the `2.8.0` release, Gateway token renewers would be run
regardless of the value of the initial auth token configured with
the Gateway client.

However, that initial value could potentially be `None`, and token
renewers are not required to support that as a value for their
`auth_token` parameter. This meant that token renewers might be
called with a value they do not support.

The `2.8.0` release fixed that issue by only calling token renewers
if the `auth_token` value was truthy. However, this introduced a
regression because the default value for auth tokens is the empty
string, which is not truthy but which *is* a valid string value
(and, as such, is expected to be supported by token renewers).

This change fixes that by calling token renewers as long as the
`auth_token` value is not `None`. This is a compromise between
the pre-2.8.0 behavior and the 2.8.0 behavior, and is believed
to be the least disruptive option.

This fixes #1339