Add prometheus metric with version information
Allows for collection of jupyter server version as a prometheus
metric, and filtering of various other metrics based on server
version.

Builds on top of conventions in other servers that expose
version (or build) info. For example, kubernetes puts out this
metric:

> kubernetes_build_info{build_date="2024-09-09T04:15:23Z", compiler="gc",
> git_commit="fa081458abcdbbc3ddfc3f99546ae5ea45aece17", git_tree_state="clean",
> git_version="v1.29.8-gke.1700", go_version="go1.22.5 X:boringcrypto",
> instance="10.128.0.50:443", job="kubernetes-apiservers", major="1", minor="29",
> platform="linux/amd64"}

Most of these parameters are not particularly relevant to us, but
version is. So this PR only adds version - future labels can be
added as needed.