Better error message when starting kernel for session.
This catches error when starting a kernel for a session (for example if
one switches the kernel for an existing notebook), and propagate the
error. Otherwise the frontend will just show "Error: uncaught
exception".

In addition this setup ruff G001 – do not use string formatting in
logging, see some of the reasons to not do that:

https://docs.astral.sh/ruff/rules/logging-string-format/

And fixes the few instances where jupyter-server does it.