Bugfix for issue 1566.
This commit provides a fix for the bug in issue #1566.
Basically we keep track of when the server side cookie
secret was generated and only trust cookies created
after that point.

When a password change occurs, we touch the cookie
secret file as a way of invalidating cookies created
before the password change. We could also generate a
new cookie secret which might be better but I am not
familiar enough with the code to know the implications.
This patch seems like the minimal way to get the bug
fixed.